Implement missing Menu entries from ChatList/View
