+++ src/renderer/components/ChatList.js
@@ -3,8 +3,7 @@ const C = require('deltachat-node/constants')
 const { ipcRenderer } = require('electron')
 const { ConversationListItem } = require('./conversations')
 
-const KeyTransferDialog = require('./dialogs/KeyTransfer')
-const DeadDropDialog = require('./dialogs/DeadDrop')
+
 
 const {
   Alignment,
@@ -22,73 +21,18 @@ const {
 class ChatList extends React.Component {
   constructor (props) {
     super(props)
-    this.state = {
-      deadDropChat: false,
-      keyTransfer: false
-    }
-    this.onDeadDropClose = this.onDeadDropClose.bind(this)
-    this.onCreateChat = this.onCreateChat.bind(this)
-    this.onCreateGroup = this.onCreateGroup.bind(this)
-    this.onCreateContact = this.onCreateContact.bind(this)
-    this.initiateKeyTransfer = this.initiateKeyTransfer.bind(this)
-    this.onKeyTransferComplete = this.onKeyTransferComplete.bind(this)
-  }
 
-  onDeadDropClose () {
-    this.setState({ deadDropChat: false })
-  }
-
-  onCreateContact () {
-    var self = this
-
-    const tx = window.translate
-
-    var onSubmit = (chatId) => {
-      if (chatId !== 0) {
-        self.props.userFeedback({ type: 'success', text: tx('contactCreateSuccess') })
-        self.props.changeScreen('ChatList')
-      }
-    }
-    this.props.changeScreen('CreateContact', { onSubmit })
-  }
-
-  onCreateChat () {
-    this.props.changeScreen('CreateChat')
-  }
-
-  onCreateGroup () {
-    this.props.changeScreen('CreateGroup')
-  }
-
-  onDeadDropClick (chat) {
-    this.setState({ deadDropChat: chat })
-  }
-
-  componentDidMount () {}
-
-  logout () {
-    ipcRenderer.send('dispatch', 'logout')
-  }
-
-  onKeyTransferComplete () {
-    this.setState({ keyTransfer: false })
-  }
-
-  initiateKeyTransfer () {
-    this.setState({ keyTransfer: true })
   }
 
   render () {
     const { deltachat, selectedChatId } = this.props
-    const { deadDropChat, keyTransfer } = this.state
 
     const tx = window.translate
 
     return (
 
       <div className="ChatList">
-        <KeyTransferDialog isOpen={keyTransfer} onClose={this.onKeyTransferComplete} />
-        <DeadDropDialog deadDropChat={deadDropChat} onClose={this.onDeadDropClose} />
+
         <div className='ConversationList'>
           {deltachat.chats.map((chat) => {
             if (!chat) return
